requirements.txt
asgiref==3.8.1
Django==4.2.20
django-rest-framework==0.1.0
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
